Fix some build issues of some TDVF test cases

+++ HBFA/UefiHostFuzzTestCasePkg/TestCase/OvmfPkg/EmuVariableFvbRuntimeDxe/TestValidateTdxCfv.c
@@ -62,6 +62,6 @@ RunTestHarness(
   // 
   // try to separate EFI lib, use stdlib function.
   // no asm code.
-  ValidateTdxCfv (TestBuffer, TestBufferSize);
+  ValidateTdxCfv (TestBuffer, (UINT32)TestBufferSize);
 
 }
